Refactor code block exit into FormatEscaper
Move code block escape logic from CommandDispatcher into FormatEscaper,
which is the existing subsystem responsible for escaping block formats
(blockquotes, lists). Add Enter-on-empty-last-line and ArrowDown-at-end
handlers alongside the existing quote and list escape handlers.

Move tests from standalone bug file into escape_format.test.js.
