+++ src/editor/contents/format_escaper.js
@@ -1,6 +1,8 @@
-import { $createParagraphNode, $getSelection, $isLineBreakNode, $isParagraphNode, $isRangeSelection, COMMAND_PRIORITY_HIGH, KEY_ENTER_COMMAND } from "lexical"
+import { $createParagraphNode, $getSelection, $isLineBreakNode, $isParagraphNode, $isRangeSelection, $isTextNode, COMMAND_PRIORITY_HIGH, COMMAND_PRIORITY_NORMAL, KEY_ARROW_DOWN_COMMAND, KEY_ENTER_COMMAND } from "lexical"
 import { $createListNode, $isListItemNode, $isListNode } from "@lexical/list"
 import { $createQuoteNode, $isQuoteNode } from "@lexical/rich-text"
+import { $isCodeNode, CodeNode } from "@lexical/code"
+import { $getNearestNodeOfType } from "@lexical/utils"
 
 export default class FormatEscaper {
   constructor(editorElement) {
@@ -14,11 +16,19 @@ export default class FormatEscaper {
       (event) => this.#handleEnterKey(event),
       COMMAND_PRIORITY_HIGH
     )
+
+    this.editor.registerCommand(
+      KEY_ARROW_DOWN_COMMAND,
+      (event) => this.#handleArrowDownInCodeBlock(event),
+      COMMAND_PRIORITY_NORMAL
+    )
   }
 
   #handleEnterKey(event) {
     const selection = $getSelection()
     if (!$isRangeSelection(selection)) return false
+
+    if (this.#handleCodeBlocks(event, selection)) return true
 
     const anchorNode = selection.anchor.getNode()
 
@@ -279,5 +289,100 @@ export default class FormatEscaper {
     this.#removeTrailingEmptyNodes(newBlockquote)
 
     newParagraph.selectStart()
+  }
+
+  // Code blocks
+
+  #handleCodeBlocks(event, selection) {
+    if (!selection.isCollapsed()) return false
+
+    const codeNode = this.#getCodeNodeFromSelection(selection)
+    if (!codeNode) return false
+
+    if (this.#isCursorOnEmptyLastLineOfCodeBlock(selection, codeNode)) {
+      event?.preventDefault()
+      this.#exitCodeBlock(codeNode)
+      return true
+    }
+
+    return false
+  }
+
+  #handleArrowDownInCodeBlock(event) {
+    const selection = $getSelection()
+    if (!$isRangeSelection(selection) || !selection.isCollapsed()) return false
+
+    const codeNode = this.#getCodeNodeFromSelection(selection)
+    if (!codeNode) return false
+
+    if (this.#isCursorOnLastLineOfCodeBlock(selection, codeNode) && !codeNode.getNextSibling()) {
+      event?.preventDefault()
+      const paragraph = $createParagraphNode()
+      codeNode.insertAfter(paragraph)
+      paragraph.selectStart()
+      return true
+    }
+
+    return false
+  }
+
+  #getCodeNodeFromSelection(selection) {
+    const anchorNode = selection.anchor.getNode()
+    return $getNearestNodeOfType(anchorNode, CodeNode) || ($isCodeNode(anchorNode) ? anchorNode : null)
+  }
+
+  #isCursorOnEmptyLastLineOfCodeBlock(selection, codeNode) {
+    const children = codeNode.getChildren()
+    if (children.length === 0) return true
+
+    const anchorNode = selection.anchor.getNode()
+    const anchorOffset = selection.anchor.offset
+
+    // Chromium: cursor on the CodeNode element after the last child (a line break)
+    if ($isCodeNode(anchorNode) && anchorOffset === children.length) {
+      return $isLineBreakNode(children[children.length - 1])
+    }
+
+    // Firefox: cursor on an empty text node that follows a line break at the end
+    if ($isTextNode(anchorNode) && anchorNode.getTextContentSize() === 0 && anchorOffset === 0) {
+      const previousSibling = anchorNode.getPreviousSibling()
+      return $isLineBreakNode(previousSibling) && anchorNode.getNextSibling() === null
+    }
+
+    return false
+  }
+
+  #isCursorOnLastLineOfCodeBlock(selection, codeNode) {
+    const anchorNode = selection.anchor.getNode()
+    const children = codeNode.getChildren()
+    if (children.length === 0) return true
+
+    const lastChild = children[children.length - 1]
+
+    if ($isCodeNode(anchorNode) && selection.anchor.offset === children.length) return true
+    if (anchorNode === lastChild) return true
+
+    const lastLineBreakIndex = children.findLastIndex(child => $isLineBreakNode(child))
+    if (lastLineBreakIndex === -1) return true
+
+    const anchorIndex = children.indexOf(anchorNode)
+    return anchorIndex > lastLineBreakIndex
+  }
+
+  #exitCodeBlock(codeNode) {
+    const children = codeNode.getChildren()
+    const lastChild = children[children.length - 1]
+
+    if ($isTextNode(lastChild) && lastChild.getTextContentSize() === 0) {
+      const previousSibling = lastChild.getPreviousSibling()
+      lastChild.remove()
+      if ($isLineBreakNode(previousSibling)) previousSibling.remove()
+    } else if ($isLineBreakNode(lastChild)) {
+      lastChild.remove()
+    }
+
+    const paragraph = $createParagraphNode()
+    codeNode.insertAfter(paragraph)
+    paragraph.selectStart()
   }
 }
